refactor(storage): Extract ConflictTracker to backends/shared.rs
ConflictTracker was defined in memory.rs but used by both memory.rs and
redb.rs, creating a cross-backend dependency. Move it to a shared module
to eliminate the `use super::memory::ConflictTracker` import from redb.
